Add Compressor and Decompressor abstract classes to refactor more common handling of streams in ByteTransformer implementations.
