+++ java/src/AdaptiveArithmeticDecompress.java
@@ -6,9 +6,8 @@
  * https://github.com/nayuki/Reference-arithmetic-coding
  */
 
-import inputstreams.InputStreamFactory;
-
-import java.io.*;
+import java.io.IOException;
+import java.io.OutputStream;
 
 
 /**
@@ -16,23 +15,14 @@ import java.io.*;
  * <p>Usage: java AdaptiveArithmeticDecompress InputFile OutputFile</p>
  * <p>This decompresses files generated by the "AdaptiveArithmeticCompress" application.</p>
  */
-public class AdaptiveArithmeticDecompress extends ByteTransformer {
+public class AdaptiveArithmeticDecompress extends DeCompressor {
 
 	public static void main(String[] args) throws IOException {
 		new AdaptiveArithmeticDecompress().commandLineMain(args);
 	}
 
 	@Override
-	public void transformStream(InputStreamFactory inputStreamFactory, OutputStream out) throws IOException {
-		try (BitInputStream in = new BitInputStream(inputStreamFactory.getStream());
-				out) {
-			decompress(in, out);
-		}
-	}
-	
-	
-	// To allow unit testing, this method is package-private instead of private.
-	static void decompress(BitInputStream in, OutputStream out) throws IOException {
+	protected void decompress(BitInputStream in, OutputStream out) throws IOException {
 		FlatFrequencyTable initFreqs = new FlatFrequencyTable(257);
 		FrequencyTable freqs = new SimpleFrequencyTable(initFreqs);
 		ArithmeticDecoder dec = new ArithmeticDecoder(32, in);
