+++ java/src/ArithmeticDecompress.java
@@ -6,9 +6,8 @@
  * https://github.com/nayuki/Reference-arithmetic-coding
  */
 
-import inputstreams.InputStreamFactory;
-
-import java.io.*;
+import java.io.IOException;
+import java.io.OutputStream;
 
 
 /**
@@ -16,23 +15,13 @@ import java.io.*;
  * <p>Usage: java ArithmeticDecompress InputFile OutputFile</p>
  * <p>This decompresses files generated by the "ArithmeticCompress" application.</p>
  */
-public class ArithmeticDecompress extends ByteTransformer {
+public class ArithmeticDecompress extends DeCompressor {
 	
 	public static void main(String[] args) throws IOException {
 		new ArithmeticDecompress().commandLineMain(args);
 	}
 	
-	@Override
-	public void transformStream(InputStreamFactory inputStreamFactory, OutputStream out) throws IOException {
-		try (BitInputStream in = new BitInputStream(inputStreamFactory.getStream());
-				out) {
-			FrequencyTable freqs = readFrequencies(in);
-			decompress(freqs, in, out);
-		}
-	}
 	
-	
-	// To allow unit testing, this method is package-private instead of private.
 	static FrequencyTable readFrequencies(BitInputStream in) throws IOException {
 		int[] freqs = new int[257];
 		for (int i = 0; i < 256; i++)
@@ -42,8 +31,9 @@ public class ArithmeticDecompress extends ByteTransformer {
 	}
 	
 	
-	// To allow unit testing, this method is package-private instead of private.
-	static void decompress(FrequencyTable freqs, BitInputStream in, OutputStream out) throws IOException {
+	@Override
+	protected void decompress(BitInputStream in, OutputStream out) throws IOException {
+		FrequencyTable freqs = readFrequencies(in);
 		ArithmeticDecoder dec = new ArithmeticDecoder(32, in);
 		while (true) {
 			int symbol = dec.read(freqs);
