+++ java/src/PpmDecompress.java
@@ -6,8 +6,6 @@
  * https://github.com/nayuki/Reference-arithmetic-coding
  */
 
-import inputstreams.InputStreamFactory;
-
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Arrays;
@@ -18,7 +16,7 @@ import java.util.Arrays;
  * <p>Usage: java PpmDecompress InputFile OutputFile</p>
  * <p>This decompresses files generated by the "PpmCompress" application.</p>
  */
-public final class PpmDecompress extends ByteTransformer {
+public final class PpmDecompress extends DeCompressor {
 	
 	// Must be at least -1 and match PpmCompress. Warning: Exponential memory usage at O(257^n).
 	private static final int MODEL_ORDER = 3;
@@ -30,16 +28,7 @@ public final class PpmDecompress extends ByteTransformer {
 	
 	
 	@Override
-	public void transformStream(InputStreamFactory inputStreamFactory, OutputStream out) throws IOException {
-		try (BitInputStream in = new BitInputStream(inputStreamFactory.getStream());
-				out) {
-			decompress(in, out);
-		}
-	}
-	
-	
-	// To allow unit testing, this method is package-private instead of private.
-	static void decompress(BitInputStream in, OutputStream out) throws IOException {
+	protected void decompress(BitInputStream in, OutputStream out) throws IOException {
 		// Set up decoder and model. In this PPM model, symbol 256 represents EOF;
 		// its frequency is 1 in the order -1 context but its frequency
 		// is 0 in all other contexts (which have non-negative order).
