+++ java/test/AdaptiveArithmeticCompressTest.java
@@ -6,24 +6,18 @@
  * https://github.com/nayuki/Reference-arithmetic-coding
  */
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-
-
 /**
  * Tests {@link AdaptiveArithmeticCompress} coupled with {@link AdaptiveArithmeticDecompress}.
  */
 public class AdaptiveArithmeticCompressTest extends ArithmeticCodingTest {
 	
 	@Override
-	protected ByteTransformer getCompressor() {
+	protected Compressor getCompressor() {
 		return new AdaptiveArithmeticCompress();
 	}
 	
 	@Override
-	protected ByteTransformer getDecompressor() {
+	protected DeCompressor getDecompressor() {
 		return new AdaptiveArithmeticDecompress();
 	}
